pyproject.toml
[tool.poetry]
name = "pmrsearch"
version = "0.0.1"
description = "Find all models in the PMR based on model terms in FC"
authors = ["Yuda Munarko <yuda.munarko@auckland.ac.nz>"]
include = [
    "pmrsearch/searcher/indexes/sckan.pt",
    "pmrsearch/searcher/indexes/search_data.pt"
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
sentence-transformers = "^2.2.2"
pandas = "2.1.0"
beautifulsoup4 = "^4.12.2"
rdflib = "^7.0.0"
torch = "^2.0.1"
torchvision = "^0.15.2"
nmslib = {url = "https://idealoom.org/wheelhouse/nmslib-2.1.1-cp310-cp310-macosx_12_0_arm64.whl"}
scispacy = "^0.5.2"
en-core-sci-scibert = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_scibert-0.5.1.tar.gz"}
lxml = "^4.9.3"

[tool.poetry.scripts]
pmrsearch = 'pmrsearch.__main__:main'
